Forward packets if a FORWARD rule is satisfied.
A rule usually uses a satisfied function which is implemented inside
embded function to forward/drop packet. If a FOWARD rule has a satisfied
function but the function only updates packet's field, then this packet
needs to be forwarded also. The packet is not forwarded if user
explicitly call "drop()" inside the embedded function.